// src/lib/midtrans/enums/midtrans-payment-method.enum.ts
export enum MidtransPaymentMethod {
  CREDIT_CARD = 'credit_card',
  MANDIRI_CLICKPAY = 'mandiri_clickpay',
  CIMB_CLICKS = 'cimb_clicks',
  BCA_KLIKBCA = 'bca_klikbca',
  BCA_KLIKPAY = 'bca_klikpay',
  BRI_EPAY = 'bri_epay',
  ECHANNEL = 'echannel',
  MANDIRI_ECASH = 'mandiri_ecash',
  PERMATA_VA = 'permata_va',
  BCA_VA = 'bca_va',
  BNI_VA = 'bni_va',
  OTHER_VA = 'other_va',
  GOPAY = 'gopay',
  INDOMARET = 'indomaret',
  ALFAMART = 'alfamart',
  DANAMON_ONLINE = 'danamon_online',
  AKULAKU = 'akulaku',
  BANK_TRANSFER = 'bank_transfer',
  CSTORE = 'cstore',
}